backend/requirements.txt
# Core FastAPI and web framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Database and ORM
sqlalchemy>=2.0.23
alembic>=1.12.1
psycopg2-binary>=2.9.9

# Data validation and settings
pydantic>=2.5.0
pydantic-settings>=2.0.3
email-validator>=2.0.0

# Authentication and security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0,<5.0.0
cryptography>=42.0.0

# Vector databases and AI
qdrant-client>=1.6.9
openai>=1.3.7
anthropic>=0.7.8
google-generativeai>=0.3.2

# HTTP client and networking
httpx>=0.25.2

# Caching and real-time features
redis>=5.0.1
websockets>=12.0
python-socketio>=5.10.0

# Document processing
PyMuPDF>=1.23.0  # More robust PDF processing with image extraction
pytesseract>=0.3.10  # OCR for scanned PDFs and images
Pillow>=10.0.0  # Image processing for OCR
python-magic>=0.4.27

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
pytest-mock>=3.11.0
httpx>=0.25.2
requests-mock>=1.11.0
factory-boy>=3.3.0
faker>=19.0.0

# Dev tooling
black>=24.3.0
ruff>=0.5.0
isort>=5.13.2
mypy>=1.10.0